Working fusion_rrf update. Takes rankings as input. No rankings should be 0 or NULL. If they are all the same value or NULL, the column is ignored. Other than that (ranking of 0 interspersed with other rankings), the behaviour is undefined (will likely assign the best score possible)

+++ src/functions/scalar/fusion_rrf/registry.cpp
@@ -5,7 +5,7 @@ namespace flockmtl {
 
 void ScalarRegistry::RegisterFusionRRF(duckdb::DatabaseInstance& db) {
     duckdb::ExtensionUtil::RegisterFunction(
-        db, duckdb::ScalarFunction("fusion_rrf", {}, duckdb::LogicalType::INTEGER, FusionRRF::Execute, nullptr,
+        db, duckdb::ScalarFunction("fusion_rrf", {}, duckdb::LogicalType::DOUBLE, FusionRRF::Execute, nullptr,
                                    nullptr, nullptr, nullptr, duckdb::LogicalType::ANY,
                                    duckdb::FunctionStability::VOLATILE, duckdb::FunctionNullHandling::SPECIAL_HANDLING));
 }
