+++ src/include/flockmtl/functions/scalar/fusion_rrf.hpp
@@ -16,7 +16,7 @@ public:
      * @param args Provided by core DuckDB when the function is executed in the CLI. Contains the arguments
      * @return A vector of rankings for the entries, which will be processed by FusionRRF::Execute
      */
-    static std::vector<int> Operation(duckdb::DataChunk& args);
+    static std::vector<double> Operation(duckdb::DataChunk& args);
     static void Execute(duckdb::DataChunk& args, duckdb::ExpressionState& state, duckdb::Vector& result);
 };
